文書順序を更新する処理を更新した。

+++ info/src/error.rs
@@ -9,7 +9,7 @@ pub enum Error {
     InvalidType,
     NotFoundDoumentElement,
     NotFoundReference(String),
-    OufOfIndex(i64),
+    OufOfIndex(usize),
     Parse(String),
 }
